Create lab4_parta_src.c
Hey Joe. Not sure if you saw Segee's comment for lab2 but he wanted separate code files for part a and for part b. I asked at the end of last class if he would like me to go back through our other labs we submitted and have two distinct files for part a and part b so I am just going back to do so now. There may be a better way to do this but I was just going to create these new files.